add package: devilspie
This is a user daemon utility that reads configuration files in lisp for
matching X11 windows and performing actions on them.

For example, I have this file:

~/.devilspie/undecorate.ds

with this contents:

(if (or (is (window_class) "qterminal") (is (window_class)
"Xfce4-terminal"))
	(undecorate)
)

With the daemon launched from my xstartup, when a window matching the
class "qtermina" or "Xfce4-terminal" is opened, it will remove window
manager decorations from it giving me a borderless window without a
title bar.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>